+++ .gitignore
@@ -8,5 +8,6 @@ app/shamir/shamir
 deploy/platform/deterlab/build
 deploy/platform/deterlab/remote/*
 deploy/platform/deterlab/deter.toml
+deploy/platform/localhost/*
 deploy/deploy
 test/
