+++ deploy/simulation/naive_single.toml
@@ -1,0 +1,7 @@
+Machines = 4
+Logservers = 0
+App = "naive"
+Suite = "ed25519"
+
+Hpn,  Rounds
+1, 1
